requirements.txt
fastapi 
uvicorn[standard]
azure-cosmos
python-dotenv
azure-appconfiguration